+++ src/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Zero2Neuro Change Log
 
 Use inverse order
+## 2025-12-03 Andrew H. Fagg Version 0.1.2
+* Added support for categorical-to-continuous feature mapping (--data_columns_categorical_to_float_direct)
 
 ## 2025-11-07 Luke Sewell  
 * Added support for tf-datasets to use pickle file reporting functionality.